requirements.txt
certifi==2022.9.24
deprecation==2.1.0
img2pdf==0.4.4
lxml==4.9.2
packaging==23.0
pdf2image==1.16.3
pikepdf==7.1.1
Pillow==9.4.0
tkinterdnd2==0.3.0
wincertstore==0.2
